+++ stream-transformation-tool-service/src/main/java/uk/gov/justice/tools/eventsourcing/transformation/service/StreamTransformer.java
@@ -6,12 +6,13 @@ import uk.gov.justice.services.eventsourcing.source.core.EventSource;
 import uk.gov.justice.services.eventsourcing.source.core.EventSourceTransformation;
 import uk.gov.justice.services.eventsourcing.source.core.EventStream;
 import uk.gov.justice.services.messaging.JsonEnvelope;
-import uk.gov.justice.tools.eventsourcing.transformation.StreamAppender;
 import uk.gov.justice.tools.eventsourcing.transformation.StreamTransformerUtil;
 import uk.gov.justice.tools.eventsourcing.transformation.api.EventTransformation;
 
+import java.util.List;
 import java.util.Set;
 import java.util.UUID;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import javax.enterprise.context.ApplicationScoped;
@@ -42,14 +43,15 @@ public class StreamTransformer {
         try {
             final EventStream stream = eventSource.getStreamById(streamId);
             final Stream<JsonEnvelope> events = stream.read();
+            final List<JsonEnvelope> originalEvents = events.collect(Collectors.toList());
 
             eventSourceTransformation.clearStream(streamId);
 
             logger.debug("Transforming events on stream {}", streamId);
 
-            final Stream<JsonEnvelope> transformedEventStream = streamTransformerUtil.transform(events, transformations);
-
-            streamAppender.appendEventsToStream(streamId, transformedEventStream);
+            try (Stream<JsonEnvelope> transformedEventStream = streamTransformerUtil.transform(originalEvents.stream(), transformations)) {
+                streamAppender.appendEventsToStream(streamId, transformedEventStream);
+            }
 
             events.close();
 
@@ -58,5 +60,6 @@ public class StreamTransformer {
         }
 
     }
+
 
 }